+++ amarillo/plugins/enhancer/configuration.py
@@ -8,7 +8,7 @@ from amarillo.app.models.Carpool import Carpool
 from amarillo.app.services import stops
 from amarillo.app.services import trips
 from amarillo.app.services.carpools import CarpoolService
-from amarillo.app.services import gtfs_generator
+from amarillo.plugins.enhancer.services import gtfs_generator
 
 from amarillo.app.configuration import configure_services
